split useAgoraCamera, useAgoraRemotes, useAgoraScreenShare into 3 separate files
